Alpha18,19: Deleted theme, ADD function
This time I added a function of inputing Z to judge whether the user inputed true
And I changed the name of this solution, so the changes you've seen are so many.
This time I added the THEME into this program, and we are able to bring this EXE without bringing any extra files!